drone_projects/requirements.txt
piexif
pillow
ultralytics
pyserial
pynmea2
flask
ffmpeg
ffmpeg-python
hyperlpr3
gmqtt
redis